src-py/pyproject.toml
[project]
name = "src-py"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "masamune", email = "kenjiii534@gmail.com" }
]
dependencies = [
    "flask>=3.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]
readme = "README.md"
requires-python = ">= 3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev_dependencies = []

[tool.hatch.metadata]
allow-direct-references = true
